requirements.txt
jsonschema==2.6.0
psycopg2-binary==2.8.6
python-dateutil==2.8.1
python-telegram-bot==13.4.1
requests==2.23.0
urllib3==1.25.9